requirements.txt
streamlit==1.25.0
SpeechRecognition==3.10.0
openai==0.28.0
pyodbc==4.0.35
faiss-cpu==1.7.4
scikit-learn==1.3.0
pandas==2.1.1
python-dotenv==1.0.0

